Move QRGenerator to the package util
