+++ app/src/main/java/com/nervousfish/nervousfish/util/QRGenerator.java
@@ -1,4 +1,4 @@
-package com.nervousfish.nervousfish;
+package com.nervousfish.nervousfish.util;
 
 import android.graphics.Bitmap;
 import android.graphics.Color;
